Linters fixes
