+++ spec/dummy/app/decorators/user_decorator.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-# Test decorator User model as a default
 class UserDecorator < Tramway::BaseDecorator
   association :posts
 end
